backend/requirements.txt
# Core LLM and AI dependencies
openai>=1.0.0
anthropic>=0.21.3
llama-cpp-python>=0.2.72
sentence-transformers>=2.2.2
langchain>=0.1.14
chromadb==0.4.24
pdfplumber>=0.10.3
PyMuPDF>=1.23.22
spacy>=3.7.2
nltk>=3.8.1
keybert>=0.8.3
yake>=0.4.8
scikit-learn>=1.4.2
rank_bm25>=0.2.2

# Google Gemini
google-generativeai>=0.3.2

# FastAPI and server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Utilities
pydantic>=2.6.4
python-dotenv>=1.0.1
requests>=2.31.0

# For testing
pytest>=8.1.1

# Document Processing
python-docx==1.1.0

# AI & ML - Enhanced Embeddings
langchain-openai==0.0.2
langchain-google-genai==0.0.6
langchain-anthropic==0.1.1

# Enhanced Embedding Models
transformers==4.36.0
torch==2.1.0

# Vector Database
faiss-cpu==1.7.4

# LLM Integrations
google-generativeai==0.3.2
anthropic==0.8.1
llama-cpp-python==0.2.20

# Utilities
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Text Processing & NLP
nltk==3.8.1
spacy==3.7.2

# Advanced Search & Retrieval
rank-bm25==0.2.2
scikit-learn==1.3.2
keybert==0.7.0
yake==0.4.8

# Semantic Analysis
gensim==4.3.2
textstat==0.7.3

# Visualization & Logging
matplotlib==3.7.2
seaborn==0.12.2 